server/content-generation-service/requirements.txt
fastapi>=0.100.0
uvicorn[standard]>=0.20.0
python-dotenv>=1.0.0
httpx>=0.25.0 # For async operations

# LLM Client Libraries
openai>=1.0.0          # For OpenAI models
anthropic>=0.7.0        # For Claude models
google-generativeai>=0.3.0  # For Gemini models
redis>=5.0.0           # For caching
pydantic>=2.0.0        # For data validation
tenacity>=8.0.0        # For retry logic
tiktoken>=0.5.0        # For token counting (OpenAI)
jinja2>=3.0.0          # For template rendering